identity/requirements.txt
# ---- Core Framework ----
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ---- Database & ORM ----
sqlalchemy[asyncio]>=2.0.28
asyncpg>=0.29.0
alembic>=1.13.1
python-dotenv>=1.0.0
greenlet>=3.0.3

# ---- Data Validation & Settings ----
pydantic>=2.6.3
pydantic-settings>=2.2.1
email-validator>=2.1.1

# ---- Utilities ----
httpx>=0.27.0
python-multipart>=0.0.9

# ---- Security (JWT & Password Hashing) ----
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==3.2.2

